Fix invoice insert and get

+++ project/src/main/java/com/miage/bigdata/daos/itemDao/column/InvoiceObjectDao.java
@@ -11,6 +11,7 @@ import com.miage.bigdata.models.column.InvoiceLine;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
     private PreparedStatement insertInvoiceStatement;
@@ -24,19 +25,25 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
 
     public InvoiceObjectDao(ColumnModelDbDao dbDao) {
         super(dbDao);
+    }
 
+    private void prepareStatements() {
+        if (insertInvoiceStatement != null) {
+            return;
+        }
 
         /**
          * Query must be prepared once
          */
-        insertInvoiceStatement = this.prepareInsertStatement( "INSERT INTO " + getDatabaseID() + "(orderId, personId, orderDate, totalPrice, orderLine) VALUES(?, ?, ?, ?, ?)");
+        insertInvoiceStatement = this.prepareInsertStatement( "INSERT INTO " + getDatabaseID() + "(orderId, personId, orderDate, totalPrice) VALUES(?, ?, ?, ?)");
         insertLineStatement = this.prepareInsertStatement("INSERT INTO invoiceLine (productId, orderId, asin, title, price, brand) VALUES(?, ?, ?, ?, ?, ?)");
         getByIdStatement = this.prepareInsertStatement("SELECT * FROM " + getDatabaseID() + " WHERE orderId = ?");
         deleteInvoiceStatement = this.prepareInsertStatement("DELETE FROM " + getDatabaseID() + " WHERE orderId = ?");
         deleteLineStatement = this.prepareInsertStatement("DELETE FROM invoiceLine WHERE productId = ?");
         getLineStatement = this.prepareInsertStatement("SELECT * FROM invoiceLine WHERE productId = ?");
         updateInvoiceStatement = this.prepareInsertStatement("UPDATE " + getDatabaseID() + " SET personId = ?, orderDate = ?, " +
-                "totalPrice = ?, orderLine = ? WHERE orderId = ?");
+                "totalPrice = ? WHERE orderId = ?");
+
         updateLineStatement = this.prepareInsertStatement("UPDATE invoiceLine SET asin = ?, title = ?, price = ?, brand = ? WHERE productId = ?");
     }
 
@@ -49,12 +56,16 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
     }
 
     private Row getLineById(String orderId) {
+        prepareStatements();
+
         BoundStatement boundStatement = new BoundStatement(getLineStatement);
         return cassandraSession.execute(boundStatement.bind(orderId)).one();
     }
 
     @Override
     public List<InvoiceItem> readAll() {
+        prepareStatements();
+
         String query = "SELECT * FROM " + getDatabaseID();
 
 
@@ -63,49 +74,58 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
         List<InvoiceItem> items = new ArrayList<>();
 
         for (Row row : rows) {
-            Row lineRow = getLineById(row.getString("orderId"));
-            items.add(instanciateItemFromRow(row, lineRow));
+            List<Row> lines = getLinesByInvoice(row.getString("orderId"));
+            items.add(instanciateItemFromRow(row, lines));
         }
 
         return items;
     }
 
+    private List<Row> getLinesByInvoice(String invoiceId) {
+        prepareStatements();
+
+        String query = "SELECT * FROM invoiceLine";
+        List<Row> rows = cassandraSession.execute(query).all();
+
+        return rows.parallelStream().filter(row -> invoiceId.equals(row.getString("orderId"))).collect(Collectors.toList());
+    }
+
     @Override
     public InvoiceItem create(InvoiceItem item) {
-        InvoiceLine invoiceLine = item.getOrderLine().get(0);
-
+        prepareStatements();
 
         BoundStatement boundStatement = new BoundStatement(insertInvoiceStatement);
         cassandraSession.execute(boundStatement.bind(item.getId(),
                 item.getPersonId(),
                 item.getOrderDate(),
-                item.getTotalPrice(),
-                invoiceLine.getId()));
+                item.getTotalPrice()));
 
-        boundStatement = new BoundStatement(insertLineStatement);
-        cassandraSession.execute(boundStatement.bind(invoiceLine.getId(),
-                item.getId(),
-                invoiceLine.getAsin(),
-                invoiceLine.getTitle(),
-                invoiceLine.getPrice(),
-                invoiceLine.getBrand()));
+        for (InvoiceLine line : item.getOrderLine()) {
+            boundStatement = new BoundStatement(insertLineStatement);
+            cassandraSession.execute(boundStatement.bind(line.getId(),
+                    item.getId(),
+                    line.getAsin(),
+                    line.getTitle(),
+                    line.getPrice(),
+                    line.getBrand()));
+        }
 
         return item;
     }
 
     @Override
     public boolean populateTable() {
+        prepareStatements();
         List<InvoiceItem> invoices = loadDataFile();
         for (InvoiceItem invoice : invoices) {
             if(invoice != null) {
-                invoice.setId(generateID());
+                //invoice.setId(generateID());
                 List<InvoiceLine> orderLines = invoice.getOrderLine();
 
                 if(orderLines.size() > 0) {
                     ColumnController columnController = new ColumnController();
                     ItemController<InvoiceLine> ilController = columnController.getItemController(InvoiceLine.class);
-                    ilController.deleteTable();
-                    ilController.createTable();
+
                     for (InvoiceLine orderLine : orderLines) {
                         orderLine.setId(generateID());
                         ilController.create(orderLine);
@@ -119,6 +139,7 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
 
     @Override
     public InvoiceItem getByID(String id) {
+        prepareStatements();
         BoundStatement boundStatement = new BoundStatement(getByIdStatement);
         Row row = cassandraSession.execute(boundStatement.bind(id)).one();
 
@@ -126,13 +147,15 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
             return null;
         }
 
-        Row lineRow = getLineById(row.getString("orderId"));
+        List<Row> lines = getLinesByInvoice(row.getString("orderId"));
 
-        return instanciateItemFromRow(row, lineRow);
+        return instanciateItemFromRow(row, lines);
     }
 
     @Override
     public boolean delete(String id) {
+        prepareStatements();
+
         Row row = getLineById(id);
 
         BoundStatement boundStatement = new BoundStatement(deleteInvoiceStatement);
@@ -146,23 +169,25 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
 
     @Override
     public InvoiceItem update(InvoiceItem item) {
-        InvoiceLine line = item.getOrderLine().get(0);
+        prepareStatements();
+
         BoundStatement boundStatement = new BoundStatement(updateInvoiceStatement);
         cassandraSession.execute(boundStatement.bind(
                 item.getPersonId(),
                 item.getOrderDate(),
                 item.getTotalPrice(),
-                item.getOrderLine().get(0).getId(),
                 item.getId()));
 
-        boundStatement = new BoundStatement(updateLineStatement);
-        cassandraSession.execute(boundStatement.bind(
-                line.getAsin(),
-                line.getTitle(),
-                line.getPrice(),
-                line.getBrand(),
-                line.getId()
-        ));
+        for (InvoiceLine line : item.getOrderLine()) {
+            boundStatement = new BoundStatement(updateLineStatement);
+            cassandraSession.execute(boundStatement.bind(
+                    line.getAsin(),
+                    line.getTitle(),
+                    line.getPrice(),
+                    line.getBrand(),
+                    line.getId()
+            ));
+        }
 
         return getByID(item.getId());
     }
@@ -178,9 +203,7 @@ public class InvoiceObjectDao extends ColumnObjectDao<InvoiceItem> {
                 "\torderId TEXT PRIMARY KEY,\n" +
                 "\tpersonId TEXT,\n" +
                 "\torderDate TIMESTAMP,\n" +
-                "\ttotalPrice DOUBLE,\n" +
-                "\torderLine TEXT\n" +
-                ");";
+                "\ttotalPrice DOUBLE\n" + ");";
 
         String createLine = "CREATE TABLE IF NOT EXISTS invoiceLine (" +
                 "productId TEXT PRIMARY KEY, " +
